refactor: purchase 패키지 기능의 테스트 코드를 purchase 패키지로 분리
