+++ src/test/java/lotto/purchase/LottoGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.purchase;
 
 import java.util.ArrayList;
 import java.util.List;
