+++ src/test/java/lotto/purchase/MyLottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.purchase;
 
 import java.util.List;
 import lotto.purchase.MyLotto;
