+++ src/test/java/lotto/purchase/PaymentExtractorTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.purchase;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
